% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{convert_between_ensembl_and_symbols}
\alias{convert_between_ensembl_and_symbols}
\title{Convert ensembl IDs for human/mouse to HGNC symbols/MGI symbols,
respectively.}
\usage{
convert_between_ensembl_and_symbols(
  genes,
  genelistSpecies = c("mouse", "human")
)
}
\arguments{
\item{genes}{Query genes as a vector of ensembl IDs.}

\item{genelistSpecies}{Either 'mouse' or 'human'.}
}
\value{
Returns a vector of gene symbols, either HGNC or MGI depending on
  selected species.
}
\description{
Allows user to convert ensembl IDs to symbols for human or mouse.
}
