% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{convert_between_species}
\alias{convert_between_species}
\title{Convert a gene list for \code{query_gene_ctd} to correct species.}
\usage{
convert_between_species(
  genes,
  genelistSpecies = c("mouse", "human"),
  ctdSpecies = c("mouse", "human")
)
}
\arguments{
\item{genes}{Query genes as a vector of MGI or HGNC symbols.}

\item{genelistSpecies}{Either 'mouse' or 'human' depending on whether MGI or
HGNC symbols are used.}

\item{ctdSpecies}{Either 'mouse', 'human', or 'both', depending on the ctd
datasets being used. Are they derived purely from mouse, human or both
species?}
}
\value{
A vector with the converted genes.
}
\description{
Allows user to convert between species dependent on species in the query gene
vector and the ctd matrices of interest.
}
