% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gene_ctd.R
\name{query_gene_ctd}
\alias{query_gene_ctd}
\title{Query genes in specificity matrices.}
\usage{
query_gene_ctd(
  genes,
  ...,
  celltypeLevel = c(1, 2),
  median_included = FALSE,
  genelistSpecies = c("mouse", "human"),
  ctdSpecies = c("mouse", "human")
)
}
\arguments{
\item{genes}{Query genes as a vector.}

\item{...}{Names of input ctds. An individual ctd, should have been generated
using \code{generate.celltype.data} from the EWCE package, and contains
mean expression and specificity matrices from a scRNAseq study.}

\item{celltypeLevel}{Either 1 or 2, depending on whether user wants to query
level 1 or 2 cell types.}

\item{median_included}{logical. If from newer version of EWCE ctd may include
median expression and specificity. Setting to TRUE will include these in
the final table.}

\item{genelistSpecies}{Either 'mouse' or 'human' depending on whether MGI or
HGNC symbols are used.}

\item{ctdSpecies}{Either 'mouse' or 'human' depending on the ctd datasets
being used. Species must be the same across all input ctds.}
}
\value{
Outputs a dataframe, with mean expression and specificity per gene
  from each study.
}
\description{
Allows user to query a vector of genes in a celltype_data file with mean
expression and specificity matrices.
}
