% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ewce.R
\name{run_ewce}
\alias{run_ewce}
\title{Perform EWCE bootstrapping across several gene lists in more than one
specificity matrix.}
\usage{
run_ewce(list_of_genes, ..., reps)
}
\arguments{
\item{list_of_genes}{List with gene lists.}

\item{...}{Names of input ctds. An individual ctd, should have been generated
using \code{generate.celltype.data} from the EWCE package, and contains
mean expression and specificity matrices from a scRNAseq study.}

\item{reps}{Number of repeats that should be performed in bootstrapping.}
}
\value{
A dataframe of EWCE results for each of the gene lists in each study.
}
\description{
Perform EWCE bootstrapping across several gene lists in more than one
specificity matrix.
}
