% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_specificity_plot.R
\name{summarise_specificity_plot}
\alias{summarise_specificity_plot}
\title{Summarised specificity plot}
\usage{
summarise_specificity_plot(filtered_specificity, fill.variable = NULL)
}
\arguments{
\item{filtered_specificity}{A dataframe that contains the columns 'CellType',
'Specificity', and 'Study', as typically outputted by
\code{query_gene_ctd}. Dataframe can contain additional columns, but must
just contain aforementioned.}

\item{flag.variable}{Default is NULL. Fill this with a column name within
your dataframe if you wish to colour by certain groups within your table.}
}
\value{
Boxplot of mean specificity values of queried gene list across cell
  types in each study.
}
\description{
This function visually summarises the mean specificity values outputted by
\code{query_gene_ctd}, across each ctd study.
}
